+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 [Unreleased](https://github.com/jshwi/docsig/compare/v0.53.2...HEAD)
 ------------------------------------------------------------------------
+### Fixed
+- param-not-equal-to-arg which should not be final catch-all
 
 [0.53.2](https://github.com/jshwi/docsig/releases/tag/v0.53.2) - 2024-05-09
 ------------------------------------------------------------------------