+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 [Unreleased](https://github.com/jshwi/docsig/compare/v0.54.0...HEAD)
 ------------------------------------------------------------------------
+### Fixed
+- issue where spelling-error could only occur alone
 
 [0.54.0](https://github.com/jshwi/docsig/releases/tag/v0.54.0) - 2024-05-21
 ------------------------------------------------------------------------